day23_CLOUDS/requirements.txt
boto3==1.24.80
python-dotenv==0.21.0
azure-storage-blob==12.13.1
dynamics365crm-python==0.2.0
requests~=2.28.1
pandas==1.5.0
msal==1.19.0
smart_open==6.2.0
flask~=2.2.2
Django~=4.1.1